Refactor LorenzSystem to use OdeSystem trait and fix Clippy issues
    - Implemented `OdeSystem` for `LorenzSystem` in `physics/chaos.rs`.
    - Removed duplicate `RungeKutta4` implementation from `chaos.rs`.
    - Refactored `LorenzSystem::step` to use `ode::RungeKutta4::step`.
    - Fixed Clippy warnings:
        - Replaced manual range checks with `RangeInclusive::contains`.
        - Replaced manual clamping with `f64::clamp`.
        - Replaced manual `is_multiple_of` checks with modulo operator (avoiding unstable API).
        - Simplified assign-op patterns.
        - Fixed needless range loop in `chaos.rs`.

    This improves composability by standardizing the ODE solver interface and reduces technical debt by adhering to idiomatic Rust patterns.

+++ math_explorer/src/physics/nuclear/mod.rs
@@ -79,10 +79,10 @@ pub mod liquid_drop {
         let z_val = atomic_number.value();
         let n_val = mass_number.value() - z_val;
 
-        let delta = if z_val % 2 == 0 && n_val % 2 == 0 {
+        let delta = if z_val as u64 % 2 == 0 && n_val as u64 % 2 == 0 {
             // Even Z, Even N
             constants::liquid_drop_constants::DELTA_COEFF * a.powf(-0.5)
-        } else if z_val % 2 != 0 && n_val % 2 != 0 {
+        } else if z_val as u64 % 2 != 0 && n_val as u64 % 2 != 0 {
             // Odd Z, Odd N
             -constants::liquid_drop_constants::DELTA_COEFF * a.powf(-0.5)
         } else {
